feature(core): Consolidate similar classes
