--- src/main/java/org/sustain/UnsupportedFeatureException.java
@@ -1,10 +1,0 @@
-package org.sustain;
-
-public class UnsupportedFeatureException extends Exception {
-    public UnsupportedFeatureException() {
-    }
-
-    public UnsupportedFeatureException(String message) {
-        super(message);
-    }
-}
