+++ src/main/java/org/sustain/datasets/controllers/IncomeController.java
@@ -1,4 +1,4 @@
-package org.sustain.census.controller;
+package org.sustain.datasets.controllers;
 
 import com.mongodb.client.MongoCollection;
 import com.mongodb.client.MongoDatabase;
