+++ src/main/java/org/sustain/datasets/controllers/PopulationController.java
@@ -1,4 +1,4 @@
-package org.sustain.census.controller;
+package org.sustain.datasets.controllers;
 
 import com.mongodb.client.FindIterable;
 import com.mongodb.client.MongoCollection;
